feat: apply Belvo layout to Playground (#456)
What
---
- Revamp Biometric Pix Playground layout to apply Belvo elements and
colors.
- Make Biometric Pix Playground layout more intuitive, describing the
necessary steps in three main sections for the flow.
- JSON fields are now beautified by default and when focusing out. This
means you can paste a JSON and it will be beautified once you click out
of the textarea.
- Add step to generate Risk Signals individually from Enrollment
creation. These two things are different steps in the product, so it
makes it easier to test.
- Rename titles and texts to match the concept of what the actions are
actually doing.
- Change default Account Tenure to be an old date to avoid any trouble
with the institutions.
- Add Belvo favicon.

Why
---
Apart from making the Playground more intuitive and easier to use, it
prepares it to be publicly available if we want, preserving the Belvo
branding.